requirements.txt
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.3.0
filelock==3.18.0
frontend==0.0.3
frozenlist==1.6.0
gptpdf==0.1.1
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
jiter==0.10.0
looseversion==1.3.0
lxml==5.4.0
multidict==6.4.4
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
openai==1.82.0
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
propcache==0.3.1
prov==2.0.1
puremagic==1.29
pydantic==2.11.5
pydantic_core==2.33.2
pydot==4.0.0
PyMuPDF==1.26.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
zipp==3.21.0
